backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic-settings==2.8.1
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
